requirements.txt
streamlit
openai
beautifulsoup4
python-dotenv
